feat: added exceptions for bad requests to GlobalExceptionHandler

+++ backend/src/main/java/xyz/streetscout/search/service/SearchService.java
@@ -5,5 +5,5 @@ import xyz.streetscout.vendor.dto.LocationDTO;
 import xyz.streetscout.vendor.dto.VendorList;
 
 public interface SearchService {
-    VendorList nearByVendors(LocationDTO location, int rangeInKm, PageRequest pageRequest);
+    VendorList nearByVendors(LocationDTO location, double rangeInKm, PageRequest pageRequest);
 }
